Fixes login variable type
If an error occurs while listing stars, the username could not be
displayed.

The README also used a previous idea version where a command to login
into GitHub and export the variables was used (which is more secure as
the TOKEN doesn't stay in the shell history) but never implemented.