Publishes Docker image upon release
The image being published tries to be as small as possible using multi
stage builds. The image supports GitHub Action labels and descriptions.

Two new `make` targets are introduced and one is changed:

* `docker`: Builds the image locally named `gh-stars`
* `docker-publish`: Publishes on Docker Hub as `augustohp/gh-stars`
* `publish`: Now executes `docker-publish` as well

Travis now logs into Docker Hub, publishing the image is automatic with
`make publish` target which already builds and publishes releases on
GitHub. The `go get` instruction got simplified so it never again need
to be changed.